예외 사항 및 세션 종료시 삭제 처리

+++ ChattingServer/Packet.h
@@ -8,7 +8,7 @@
 class CPacket
 {
 public:
-	enum class en_PACKET : UINT8
+	enum class en_PACKET : UINT32
 	{
 		eBUFFER_DEFAULT = 1400		// 패킷의 기본 버퍼 사이즈.
 	};
@@ -18,7 +18,7 @@ public:
 	//
 	// Return:
 	//////////////////////////////////////////////////////////////////////////
-	explicit CPacket(int iBufferSize = static_cast<UINT8>(en_PACKET::eBUFFER_DEFAULT));
+	explicit CPacket(int iBufferSize = static_cast<UINT32>(en_PACKET::eBUFFER_DEFAULT));
 	virtual ~CPacket();
 
 	//////////////////////////////////////////////////////////////////////////
